requirements.txt
boto3
pyyaml
tqdm
google-api-python-client
google-auth
google-cloud-container
